Stop adding whitespace when parsing our html on page morph.
As shown by @shepmaster on #487, we were adding extra whitespace
to the DOM when parsing. That was causing morphdom to do several
replaces on the first page load since it was comparing a page without
the whitespace and the "new" page with all the whitespaces added.

The solution was just saying for Nokogiri to use the DEFAULT_HTML
but without doing the FORMAT part.

https://github.com/hopsoft/stimulus_reflex/issues/487#issuecomment-816051404

Closes #487